AutoHyperlinks.framework will now compile if Max OS X 10.5 and 64-bit build options are set.

+++ Frameworks/AutoHyperlinks Framework/Source/AHHyperlinkScanner.h
@@ -27,14 +27,14 @@
 
 #import "AHLinkLexer.h"
 
-extern int AHleng;
-extern int AHlex();
+extern long AHleng;
+extern long AHlex();
 typedef struct AH_buffer_state *AH_BUFFER_STATE;
 void AH_switch_to_buffer(AH_BUFFER_STATE);
 AH_BUFFER_STATE AH_scan_string (const char *);
 void AH_delete_buffer(AH_BUFFER_STATE);
 
-extern unsigned int AHStringOffset;
+extern unsigned long AHStringOffset;
 
 @class AHMarkedHyperlink;
